+++ CHANGES.rst
@@ -43,6 +43,9 @@ New Features
 
 - ``astropy.io.fits``
 
+  - Included a new command-line script called ``fitshead`` to display the
+    header(s) of a FITS file from the command line. [#2092]
+
 - ``astropy.io.misc``
 
 - ``astropy.io.registry``